[F] Extract layout into separate component
