+++ src/components/codeOfConductText/CodeOfConductText.tsx
@@ -1,54 +1,43 @@
 import React from "react"
 
 export const CodeOfConductText: React.FC = () => {
-  return (
-    <>
-      <div>
-        <h2>{"Code of Conduct"}</h2>
-        <ul>
-          <li>
-            <b>
-              {"🫱🏼‍🫲🏾 Respect: "}
-            </b>
-            {"We respect each other as peers, colleagues, and collaborators."}
-          </li>
+	return (
+		<>
+			<div>
+				<ul>
+					<li>
+						<b>{"🫱🏼‍🫲🏾 Respect: "}</b>
+						{"We respect each other as peers, colleagues, and collaborators."}
+					</li>
 
-          <li>
-            <b>
-              {"💆🏽‍♀️ Self-care: "}
-            </b>
-            {"We allow ourselves to celebrate, and allow ourselves to fail."}
-          </li>
+					<li>
+						<b>{"💆🏽‍♀️ Self-care: "}</b>
+						{"We allow ourselves to celebrate, and allow ourselves to fail."}
+					</li>
 
-          <li>
-            <b>
-              {"🫂 Community: "}
-            </b>
-            {"We are invested in helping each other and sharing knowledge."}
-          </li>
+					<li>
+						<b>{"🫂 Community: "}</b>
+						{"We are invested in helping each other and sharing knowledge."}
+					</li>
 
-          <li>
-            <b>
-              {"🌱 Growth: "}
-            </b>
-            {"We offer honest and constructive feedback, not judgment."}
-          </li>
+					<li>
+						<b>{"🌱 Growth: "}</b>
+						{"We offer honest and constructive feedback, not judgment."}
+					</li>
 
-          <li>
-            <b>
-              {"🔦 Dedication: "}
-            </b>
-            {"We work hard, but we don’t presume what that means."}
-          </li>
+					<li>
+						<b>{"🔦 Dedication: "}</b>
+						{"We work hard, but we don’t presume what that means."}
+					</li>
 
-          <li>
-            <b>
-              {"🦾 Integrity: "}
-            </b>
-            {"We bring our lessons to our interviews, not our interviews to our lessons."}
-          </li>
-        </ul>
-      </div>
-    </>
-  )
+					<li>
+						<b>{"🦾 Integrity: "}</b>
+						{
+							"We bring our lessons to our interviews, not our interviews to our lessons."
+						}
+					</li>
+				</ul>
+			</div>
+		</>
+	)
 }
