+++ src/components/homePageHeader/HomePageHeader.tsx
@@ -1,11 +1,11 @@
 import React from "react"
-import styles from "./header.module.css"
+import styles from "./homePageHeader.module.css"
 
 type PropsType = {
 	subTitle?: string
 }
 
-export const Header: React.FC<PropsType> = ({ subTitle }) => {
+export const HomePageHeader: React.FC<PropsType> = ({ subTitle }) => {
 	return (
 		<div className={styles.wrapper}>
 			<h1 className={styles.title}>{"Empathetech"}</h1>
