+++ src/components/homePageHeader/homePageHeader.module.css
@@ -7,6 +7,7 @@
 	text-align: center;
 	font-size: 50px;
 	font-weight: var(--font-weight-bold);
+	padding-bottom: 0;
 }
 
 @media screen and (max-width: 550px) {
