+++ src/components/index.ts
@@ -3,7 +3,8 @@ export { Button } from "./button/Button"
 export { Calendar } from "./calendar/Calendar"
 export { CodeOfConductText } from "./codeOfConductText/CodeOfConductText"
 export { Footer } from "./footer/Footer"
-export { Header } from "./header/Header"
+export { HomePageHeader } from "./homePageHeader/HomePageHeader"
+export { Layout } from "./layout/Layout"
 export { LinkButton } from "./linkButton/LinkButton"
 export { NavigationBar } from "./navigationBar/NavigationBar"
 export { TextBlock } from "./textBlock/TextBlock"
