+++ src/components/layout/layout.module.css
@@ -1,0 +1,6 @@
+.main {
+	display: flex;
+	flex-direction: column;
+	align-items: center;
+	min-height: calc(100vh - 200px);
+}
