+++ src/pages/404.tsx
@@ -11,9 +11,7 @@ export default function ErrorPage() {
 				<meta name="viewport" content="width=device-width, initial-scale=1" />
 				<link rel="icon" href="/favicon.ico" />
 			</Head>
-			<main className={styles.main}>
-				<h1>{"Rats! This page cannot be found. 🐀"}</h1>
-			</main>
+			<h1>{"Rats! This page cannot be found. 🐀"}</h1>
 		</>
 	)
 }
