+++ src/pages/events.tsx
@@ -12,10 +12,8 @@ export default function Example() {
 				<meta name="viewport" content="width=device-width, initial-scale=1" />
 				<link rel="icon" href="/favicon.ico" />
 			</Head>
-			<main className={styles.main}>
-				<h1>{"Events"}</h1>
-				<Calendar />
-			</main>
+			<h1>{"Events"}</h1>
+			<Calendar />
 		</>
 	)
 }
