+++ src/pages/index.tsx
@@ -2,7 +2,7 @@ import Head from "next/head"
 import styles from "@/styles/page.module.css"
 import {
 	AboutEmpathetechText,
-	Header,
+	HomePageHeader,
 	LinkButton,
 	TextBlock,
 } from "@/components"
@@ -20,25 +20,21 @@ export default function Home() {
 				<meta name="viewport" content="width=device-width, initial-scale=1" />
 				<link rel="icon" href="/favicon.ico" />
 			</Head>
-			<main className={styles.main}>
-				<Header subTitle={TAGLINE} />
-				<div className={styles.buttonWrapper}>
-					<LinkButton
-						link="https://join.slack.com/t/empathetech/shared_invite/zt-1vocficwm-UywQkATY27UXIce8eU3onA"
-						isExternal
-					>
-						{"Join our Slack"}
-						<div className={styles.iconWrapper}>
-							<SlackIcon height={28} width={28} />
-						</div>
-					</LinkButton>
-				</div>
-
-				<TextBlock>
-					<AboutEmpathetechText />
-				</TextBlock>
-			</main>
-
+			<HomePageHeader subTitle={TAGLINE} />
+			<div className={styles.buttonWrapper}>
+				<LinkButton
+					link="https://join.slack.com/t/empathetech/shared_invite/zt-1vocficwm-UywQkATY27UXIce8eU3onA"
+					isExternal
+				>
+					{"Join our Slack"}
+					<div className={styles.iconWrapper}>
+						<SlackIcon height={28} width={28} />
+					</div>
+				</LinkButton>
+			</div>
+			<TextBlock>
+				<AboutEmpathetechText />
+			</TextBlock>
 		</>
 	)
 }
