+++ src/styles/page.module.css
@@ -1,11 +1,3 @@
-.main {
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-	justify-content: center;
-	min-height: calc(100vh - 200px);
-}
-
 .buttonWrapper {
 	padding-bottom: var(--size-m);
 }
@@ -14,11 +6,4 @@
 	display: flex;
 	align-items: center;
 	padding-left: var(--size-xs);
-}
-
-.navigation {
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-	justify-content: center;
 }
